# openapi.yaml
openapi: 3.0.3
info:
  title: FIXO PASS API
  description: >
    API de compartilhamento rápido e autorizado de dados entre usuários e empresas.
    O FIXO PASS não é um documento digital e não substitui RG, CPF, CNH ou os
    sistemas das empresas — ele apenas entrega, mediante autorização explícita
    do usuário, os dados que a empresa configurou como necessários.
  version: 0.1.0
servers:
  - url: http://localhost:3000
    description: Ambiente local de desenvolvimento
components:
  securitySchemes:
    CompanyApiKey:
      type: apiKey
      in: header
      name: X-API-KEY
      description: Chave de API emitida para a empresa/ERP integrado.
    UserSession:
      type: apiKey
      in: header
      name: X-USER-ID
      description: MVP — identificador do usuário logado no app (substituir por JWT em produção).
    CompanyPanelSession:
      type: apiKey
      in: header
      name: X-COMPANY-ID
      description: MVP — identificador da empresa logada no painel web (diferente da X-API-KEY do ERP).
    AdminSecret:
      type: apiKey
      in: header
      name: X-ADMIN-SECRET
      description: Segredo compartilhado do Super Admin (variável de ambiente ADMIN_SECRET) — acesso a todas as empresas.
    PaymentWebhookSecret:
      type: apiKey
      in: header
      name: X-WEBHOOK-SECRET
      description: Segredo compartilhado do webhook de pagamento (variável de ambiente PAYMENT_WEBHOOK_SECRET) — NÃO é verificação de assinatura oficial do gateway, ver comentário em src/routes/webhooks.ts.
    IntegrationBasicAuth:
      type: http
      scheme: basic
      description: >
        client_id/client_secret de uma Integração do canal WEB/API (ver
        POST /companies/me/integracoes), enviados como Authorization: Basic
        base64(clientId:clientSecret). Usado apenas server-to-server pelo
        backend do sistema externo — nunca pelo navegador do cliente final.
  schemas:
    Campo:
      type: string
      enum: [NOME, TELEFONE, EMAIL, CPF, RG, DATA_NASCIMENTO, ENDERECO, FOTO]
...
    ShareInput:
      type: object
      required: [solicitacaoId, aprovar]
      properties:
        solicitacaoId:
          type: string
          format: uuid
        aprovar:
          type: boolean
        camposLiberados:
          type: array
          items:
            $ref: '#/components/schemas/Campo'
    ShareOutput:
      type: object
      properties:
        status:
          type: string
          enum: [APROVADA, NEGADA]
        dados:
          type: object
          additionalProperties: true
          example:
            nome: "Maria Silva"
            telefone: "+55 19 99999-0000"
            endereco: "Rua das Flores, 123"
        accessCount:
          type: integer
          description: "Nº de pareamentos aprovados do usuário com essa empresa (inclui este). Ausente quando status=NEGADA."
        lastAccessedAt:
          type: string
          format: date-time
          description: "Ausente quando status=NEGADA."
    CustomerDataOutput:
      type: object
      properties:
        dados:
          type: object
          additionalProperties: true
        accessCount:
          type: integer
          description: Nº de pareamentos aprovados desse usuário com essa empresa
        lastAccessedAt:
          type: string
          format: date-time
...
paths:
...
  /customer/share:
    post:
      summary: Usuário aprova/nega a solicitação e o FIXO PASS envia os dados autorizados
      security:
        - UserSession: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ShareInput'
      responses:
        '200':
          description: Solicitação resolvida
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShareOutput'
        '400':
          description: Dados inválidos
        '404':
          description: Solicitação não encontrada
        '409':
          description: Solicitação já havia sido resolvida
  /customer/{id}:
    get:
      summary: Empresa consulta os dados já autorizados de um cliente recorrente
      security:
        - CompanyApiKey: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Dados autorizados do cliente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerDataOutput'
        '401':
          description: API Key ausente ou inválida
        '403':
          description: Nenhuma autorização ativa para este usuário
        '404':
          description: Usuário não encontrado
  /integrations/authorization-requests:
    post:
      summary: "[Canal WEB/API] Sistema externo (ex.: cardápio online) inicia um compartilhamento"
      description: >
        Chamado pelo BACKEND do sistema externo (nunca pelo navegador do
        cliente). Equivalente do POST /auth/request para o canal Web — a
        diferença é que aqui ainda não se sabe qual usuário vai autorizar;
        isso só é resolvido quando a pessoa abre a authorizationUrl e loga
        no FIXO PASS.
      security:
        - IntegrationBasicAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [redirectUri]
              properties:
                purpose:
                  type: string
                  description: Finalidade em texto livre, mostrada na tela de consentimento.
                requestedFields:
                  type: array
                  description: Opcional — se omitido, usa os campos configurados pela empresa em PUT /companies/me/campos-solicitados. Precisa ser subconjunto desses campos.
                  items:
                    $ref: '#/components/schemas/Campo'
                redirectUri:
                  type: string
                  format: uri
                  description: Precisa estar entre as redirectUris cadastradas na integração.
      responses:
        '201':
          description: Solicitação criada
          content:
            application/json:
              schema:
                type: object
                properties:
                  requestId:
                    type: string
                    format: uuid
                  authorizationUrl:
                    type: string
                    format: uri
                  expiresAt:
                    type: string
                    format: date-time
        '400':
          description: Dados inválidos, redirectUri não cadastrada ou campo não permitido para esta integração
        '401':
          description: Credenciais de integração ausentes ou inválidas
        '402':
          description: Assinatura da empresa vencida
  /integrations/authorize/{requestId}:
    get:
      summary: "[Canal WEB/API] Tela de consentimento consulta a solicitação"
      description: Chamado pela tela de consentimento (app/authorize.html) já com o usuário logado no FIXO PASS.
      security:
        - UserSession: []
      parameters:
        - name: requestId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Detalhes para exibir na tela de consentimento (empresa, finalidade, campos pedidos, status)
        '401':
          description: Usuário não autenticado
        '403':
          description: Solicitação já associada a outro usuário
        '404':
          description: Solicitação não encontrada
    post:
      summary: "[Canal WEB/API] Usuário aprova ou nega a solicitação"
      description: >
        Reaproveita o mesmo núcleo de POST /customer/share — a diferença é
        que NUNCA devolve os dados pro navegador, só um authorization_code
        opaco, de uso único e validade curta (2 minutos), que o backend do
        sistema externo troca em POST /integrations/token.
      security:
        - UserSession: []
      parameters:
        - name: requestId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [aprovar]
              properties:
                aprovar:
                  type: boolean
                camposLiberados:
                  type: array
                  items:
                    $ref: '#/components/schemas/Campo'
      responses:
        '200':
          description: "Solicitação resolvida — status APROVADA inclui `code`; NEGADA não."
        '400':
          description: Dados inválidos
        '401':
          description: Usuário não autenticado
        '403':
          description: Solicitação não pertence a este usuário
        '404':
          description: Solicitação não encontrada
        '409':
          description: Solicitação já havia sido resolvida (ou expirou)
  /integrations/token:
    post:
      summary: "[Canal WEB/API] Backend do sistema externo troca o authorization_code pelos dados autorizados"
      security:
        - IntegrationBasicAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code, redirectUri]
              properties:
                code:
                  type: string
                redirectUri:
                  type: string
                  format: uri
                  description: Precisa ser idêntica à usada em POST /integrations/authorization-requests.
      responses:
        '200':
          description: Dados autorizados — SOMENTE os campos que o usuário liberou
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    description: Chaves em minúsculo (nome, telefone, endereco...), só as autorizadas.
        '400':
          description: Código inválido, expirado ou já utilizado; ou redirectUri não confere
        '401':
          description: Credenciais de integração ausentes ou inválidas
        '403':
          description: Este código não pertence a esta integração (isolamento multi-tenant)
...